+++ server.go
@@ -233,8 +233,7 @@ func (s *server) runTcpEventLoop(newSession NewSessionCallback) {
 				return
 			}
 			if delay != 0 {
-				// time.Sleep(delay)
-				<-wheel.After(delay)
+				time.Sleep(delay)
 			}
 			client, err = s.accept(newSession)
 			if err != nil {