+++ change_log.md
@@ -14,6 +14,11 @@
 ## develop history ##
 ---
 
+- 2019/09/05
+    > Feature
+    * add writev
+    * version: v1.1.2
+   
 - 2018/10/16
     > Feature
     * add rpc/micro CallOneway/AsyncCall